cloud/functions/requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.9.3
bigquery==0.0.12
BigQuery-Python==1.15.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
Cython==0.29.22
dacktool==0.0.7
DateTime==4.3
dbstream==0.0.30
feedparser==6.0.2
firebase-admin==4.5.2
Flask==1.1.2
gensim==3.8.3
google-api-core==1.26.1
google-api-python-client==1.7.11
google-auth==1.27.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.3
google-cloud-bigquery==2.12.0
google-cloud-bigquery-storage==2.3.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
googleauthentication==0.0.15
grpcio==1.36.1
httplib2==0.19.0
idna==2.10
itsdangerous==1.1.0
Janome==0.4.1
Jinja2==2.11.3
joblib==1.0.1
libcst==0.3.17
MarkupSafe==1.1.1
metadata-parser==0.10.4
msgpack==1.0.2
mypy-extensions==0.4.3
nltk==3.2.4
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
paramiko==2.7.2
proto-plus==1.17.0
protobuf==3.15.6
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
smart-open==4.2.0
soupsieve==2.2
sphinx-rtd-theme==0.2.4
sshtunnel==0.1.5
threadpoolctl==2.1.0
tqdm==4.59.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
zope.interface==5.2.0
